requirements.txt
django
djangorestframework
psycopg2
langdetect
django-extensions
factory_boy
google-api-python-client
python-dotenv

